fix(positions): keep inferred heat level strongly typed

+++ lib/features/positions/data/position_repository_asset.dart
@@ -59,7 +59,7 @@ class PositionRepositoryAsset implements PositionRepository {
               .join(' ');
     final heatMin = zone == PositionZone.wild
         ? 4
-        : 1 + (index ~/ 48).clamp(0, 3);
+        : 1 + (index ~/ 48).clamp(0, 3).toInt();
     return IntimacyPosition(
       id: 'asset_pos_${index + 1}',
       zone: zone,
